+++ src/compas_timber/connections/plate_butt_joint.py
@@ -9,37 +9,44 @@ class PlateButtJoint(PlateJoint):
     """Creates a plate-to-plate butt-joint connection."""
 
     @property
-    def __data__(self):
-        data = super(PlateJoint, self).__data__
-        data["main_plate_guid"] = self._main_plate_guid
-        data["cross_plate_guid"] = self._cross_plate_guid
-        data["topology"] = self.topology
-        data["main_segment_index"] = self.main_segment_index
-        data["cross_segment_index"] = self.cross_segment_index
-        return data
-
-    def __init__(self, main_plate, cross_plate, topology, main_segment_index, cross_segment_index, **kwargs):
-        super(PlateButtJoint, self).__init__(main_plate, cross_plate, topology, main_segment_index, cross_segment_index, **kwargs)
-
-    @property
     def main_plate(self):
         """Return the main plate."""
         return self.plate_a
+
+    @main_plate.setter
+    def main_plate(self, value):
+        """Set the main plate."""
+        self.plate_a = value
 
     @property
     def cross_plate(self):
         """Return the cross plate."""
         return self.plate_b
 
+    @cross_plate.setter
+    def cross_plate(self, value):
+        """Set the cross plate."""
+        self.plate_b = value
+
     @property
     def main_segment_index(self):
         """Return the index of the segment in the main plate outline."""
         return self.a_segment_index
 
+    @main_segment_index.setter
+    def main_segment_index(self, value):
+        """Set the index of the segment in the main plate outline."""
+        self.a_segment_index = value
+
     @property
     def cross_segment_index(self):
-        """Return the index of the segment in the main plate outline."""
+        """Return the index of the segment in the cross plate outline."""
         return self.b_segment_index
+
+    @cross_segment_index.setter
+    def cross_segment_index(self, value):
+        """Set the index of the segment in the cross plate outline."""
+        self.b_segment_index = value
 
     @property
     def _main_plate_guid(self):
@@ -113,19 +120,6 @@ class PlateLButtJoint(PlateButtJoint):
 
     SUPPORTED_TOPOLOGY = JointTopology.TOPO_EDGE_EDGE
 
-    @property
-    def __data__(self):
-        data = super(PlateLButtJoint, self).__data__
-        data["main_plate_guid"] = self._main_plate_guid
-        data["cross_plate_guid"] = self._cross_plate_guid
-        data["topology"] = self.topology
-        data["main_segment_index"] = self.main_segment_index
-        data["cross_segment_index"] = self.cross_segment_index
-        return data
-
-    def __init__(self, main_plate, cross_plate, topology, main_segment_index, cross_segment_index, **kwargs):
-        super(PlateLButtJoint, self).__init__(main_plate, cross_plate, topology, main_segment_index, cross_segment_index, **kwargs)
-
     def __repr__(self):
         return "PlateLButtJoint({0}, {1}, {2})".format(self.main_plate, self.cross_plate, JointTopology.get_name(self.topology))
 
@@ -146,18 +140,6 @@ class PlateTButtJoint(PlateButtJoint):
     """Creates a plate-to-plate butt-joint connection."""
 
     SUPPORTED_TOPOLOGY = JointTopology.TOPO_EDGE_FACE
-
-    @property
-    def __data__(self):
-        data = super(PlateTButtJoint, self).__data__
-        data["main_plate_guid"] = self._main_plate_guid
-        data["cross_plate_guid"] = self._cross_plate_guid
-        data["topology"] = self.topology
-        data["main_segment_index"] = self.main_segment_index
-        return data
-
-    def __init__(self, main_plate, cross_plate, topology, main_segment_index, **kwargs):
-        super(PlateTButtJoint, self).__init__(main_plate, cross_plate, topology, main_segment_index, **kwargs)
 
     def __repr__(self):
         return "PlateTButtJoint({0}, {1}, {2})".format(self.main_plate, self.cross_plate, JointTopology.get_name(self.topology))